refactor: 클래스의 역할과 의도에 맞게 LottoNumberGenerator가 아닌 LottoNumberAutoGenerator로 변경 - LottoNumberAutoGenerator의 기본 구현 클래스인 DefaultLottoNumberAutoGenerator로 이름 변경 - 테스트 코드 및 애플리케이션 코드 대응 작업
