--- src/main/kotlin/lotto/domain/strategy/LottoNumberGenerator.kt
@@ -1,8 +1,0 @@
-package lotto.domain.strategy
-
-import lotto.domain.LottoNumbers
-
-interface LottoNumberGenerator {
-
-    fun generate(): LottoNumbers
-}
