+++ src/test/kotlin/lotto/domain/strategy/MockLottoNumberAutoGenerator.kt
@@ -3,7 +3,7 @@ package lotto.domain.strategy
 import lotto.domain.LottoNumbers
 import lotto.fixture.LottoNumbersFixture
 
-object MockLottoNumberGenerator : LottoNumberGenerator {
+object MockLottoNumberAutoGenerator : LottoNumberAutoGenerator {
 
     override fun generate(): LottoNumbers {
         return LottoNumbersFixture.LOTTO_NUMBER_SET_WINNING
